+++ src/babashka/main.clj
@@ -2,31 +2,39 @@
   {:no-doc true}
   (:require
    [babashka.impl.async :refer [async-namespace async-protocols-namespace]]
+   [babashka.impl.bencode :refer [bencode-namespace]]
    [babashka.impl.cheshire :refer [cheshire-core-namespace]]
    [babashka.impl.classes :as classes]
    [babashka.impl.classpath :as cp]
    [babashka.impl.clojure.core :refer [core-extras]]
    [babashka.impl.clojure.java.io :refer [io-namespace]]
    [babashka.impl.clojure.java.shell :refer [shell-namespace]]
-   [babashka.impl.clojure.main :refer [demunge]]
-   [babashka.impl.clojure.stacktrace :refer [stacktrace-namespace print-stack-trace]]
+   [babashka.impl.clojure.main :as clojure-main :refer [demunge]]
+   [babashka.impl.clojure.pprint :refer [pprint-namespace]]
+   [babashka.impl.clojure.stacktrace :refer [stacktrace-namespace]]
    [babashka.impl.common :as common]
    [babashka.impl.csv :as csv]
+   [babashka.impl.curl :refer [curl-namespace]]
+   [babashka.impl.nrepl-server :as nrepl-server]
    [babashka.impl.pipe-signal-handler :refer [handle-pipe! pipe-signal-received?]]
    [babashka.impl.repl :as repl]
+   [babashka.impl.sigint-handler :as sigint-handler]
    [babashka.impl.socket-repl :as socket-repl]
    [babashka.impl.test :as t]
    [babashka.impl.tools.cli :refer [tools-cli-namespace]]
    [babashka.impl.xml :as xml]
+   [babashka.impl.transit :refer [transit-namespace]]
+   [babashka.impl.yaml :refer [yaml-namespace]]
    [babashka.wait :as wait]
    [clojure.edn :as edn]
    [clojure.java.io :as io]
+   [clojure.stacktrace :refer [print-stack-trace]]
    [clojure.string :as str]
-   [fipp.edn :as fipp]
    [sci.addons :as addons]
    [sci.core :as sci]
    [sci.impl.interpreter :refer [eval-string*]]
    [sci.impl.opts :as sci-opts]
+   [sci.impl.types :as sci-types]
    [sci.impl.unrestrict :refer [*unrestricted*]]
    [sci.impl.vars :as vars])
   (:gen-class))
@@ -110,7 +118,14 @@
                      (let [options (next options)]
                        (recur (next options)
                               (assoc opts-map
-                                     :socket-repl (first options))))
+                                     :socket-repl (or (first options)
+                                                      "1666"))))
+                     ("--nrepl-server")
+                     (let [options (next options)]
+                       (recur (next options)
+                              (assoc opts-map
+                                     :nrepl (or (first options)
+                                                      "1667"))))
                      ("--eval", "-e")
                      (let [options (next options)]
                        (recur (next options)
@@ -154,7 +169,7 @@
 (def usage-string "Usage: bb [ -i | -I ] [ -o | -O ] [ --stream ] [--verbose]
           [ ( --classpath | -cp ) <cp> ] [ --uberscript <file> ]
           [ ( --main | -m ) <main-namespace> | -e <expression> | -f <file> |
-            --repl | --socket-repl [<host>:]<port> ]
+            --repl | --socket-repl [<host>:]<port> | --nrepl-server [<host>:]<port> ]
           [ arg* ]")
 (defn print-usage []
   (println usage-string))
@@ -184,6 +199,7 @@
   -m, --main <ns>     Call the -main function from namespace with args.
   --repl              Start REPL. Use rlwrap for history.
   --socket-repl       Start socket REPL. Specify port (e.g. 1666) or host and port separated by colon (e.g. 127.0.0.1:1666).
+  --nrepl-server      Start nREPL server. Specify port (e.g. 1667) or host and port separated by colon (e.g. 127.0.0.1:1667).
   --time              Print execution time before exiting.
   --                  Stop parsing args and pass everything after -- to *command-line-args*
 
@@ -198,25 +214,26 @@ Everything after that is bound to *command-line-args*."))
         (str/replace x #"^#!.*" ""))
       (throw (Exception. (str "File does not exist: " file))))))
 
-(defn read-edn []
-  (edn/read {;;:readers *data-readers*
-             :eof ::EOF} *in*))
-
 (def reflection-var (sci/new-dynamic-var '*warn-on-reflection* false))
 
 (defn load-file* [sci-ctx f]
   (let [f (io/file f)
-        s (slurp f)]
+        s (slurp f)
+        prev-ns @vars/current-ns]
     (sci/with-bindings {vars/current-file (.getCanonicalPath f)}
-      (eval-string* sci-ctx s))))
-
-(defn eval* [sci-ctx form]
-  (eval-string* sci-ctx (pr-str form)))
+      (try
+        (eval-string* sci-ctx s)
+        (finally (sci-types/setVal vars/current-ns prev-ns))))))
 
 (defn start-socket-repl! [address ctx]
   (socket-repl/start-repl! address ctx)
   ;; hang until SIGINT
   @(promise))
+
+(defn start-nrepl! [address ctx]
+  (nrepl-server/start-server! ctx address)
+  ;; hang until SIGINT
+  #_@(promise))
 
 (defn exit [n]
   (throw (ex-info "" {:bb/exit-code n})))
@@ -231,7 +248,11 @@ Everything after that is bound to *command-line-args*."))
     async clojure.core.async
     csv clojure.data.csv
     json cheshire.core
-    xml clojure.data.xml})
+    xml clojure.data.xml
+    yaml clj-yaml.core
+    curl babashka.curl
+    transit cognitect.transit
+    bencode bencode.core})
 
 (def cp-state (atom nil))
 
@@ -257,12 +278,17 @@ Everything after that is bound to *command-line-args*."))
    'clojure.data.csv csv/csv-namespace
    'cheshire.core cheshire-core-namespace
    'clojure.stacktrace stacktrace-namespace
-   'clojure.main {'demunge demunge}
+   'clojure.main {'demunge demunge
+                  'repl-requires clojure-main/repl-requires}
    'clojure.repl {'demunge demunge}
    'clojure.test t/clojure-test-namespace
    'babashka.classpath {'add-classpath add-classpath*}
-   'clojure.pprint {'pprint fipp/pprint}
-   'clojure.data.xml xml/xml-namespace})
+   'clojure.data.xml xml/xml-namespace
+   'clj-yaml.core yaml-namespace
+   'clojure.pprint pprint-namespace
+   'babashka.curl curl-namespace
+   'cognitect.transit transit-namespace
+   'bencode.core bencode-namespace})
 
 (def bindings
   {'java.lang.System/exit exit ;; override exit, so we have more control
@@ -284,6 +310,7 @@ Everything after that is bound to *command-line-args*."))
 (defn main
   [& args]
   (handle-pipe!)
+  (sigint-handler/handle-sigint!)
   #_(binding [*out* *err*]
       (prn "M" (meta (get bindings 'future))))
   (binding [*unrestricted* true]
@@ -293,16 +320,18 @@ Everything after that is bound to *command-line-args*."))
             {:keys [:version :shell-in :edn-in :shell-out :edn-out
                     :help? :file :command-line-args
                     :expressions :stream? :time?
-                    :repl :socket-repl
+                    :repl :socket-repl :nrepl
                     :verbose? :classpath
                     :main :uberscript] :as _opts}
             (parse-opts args)
+            _ (when main (System/setProperty "babashka.main" main))
             read-next (fn [*in*]
                         (if (pipe-signal-received?)
                           ::EOF
                           (if stream?
                             (if shell-in (or (read-line) ::EOF)
-                                (read-edn))
+                                (edn/read {;;:readers *data-readers*
+                                           :eof ::EOF} *in*))
                             (delay (cond shell-in
                                          (shell-seq *in*)
                                          edn-in
@@ -337,6 +366,7 @@ Everything after that is bound to *command-line-args*."))
                  :imports '{ArithmeticException java.lang.ArithmeticException
                             AssertionError java.lang.AssertionError
                             Boolean java.lang.Boolean
+                            Byte java.lang.Byte
                             Class java.lang.Class
                             Double java.lang.Double
                             Exception java.lang.Exception
@@ -345,7 +375,10 @@ Everything after that is bound to *command-line-args*."))
                             File java.io.File
                             Long java.lang.Long
                             Math java.lang.Math
+                            NumberFormatException java.lang.NumberFormatException
                             Object java.lang.Object
+                            Runtime java.lang.Runtime
+                            RuntimeException java.lang.RuntimeException
                             ProcessBuilder java.lang.ProcessBuilder
                             String java.lang.String
                             StringBuilder java.lang.StringBuilder
@@ -357,17 +390,20 @@ Everything after that is bound to *command-line-args*."))
             ctx (addons/future ctx)
             sci-ctx (sci-opts/init ctx)
             _ (vreset! common/ctx sci-ctx)
+            input-var (sci/new-dynamic-var '*input* nil)
             _ (swap! (:env sci-ctx)
                      (fn [env]
-                       (update-in env [:namespaces 'clojure.core] assoc
-                                  'eval #(eval* sci-ctx %)
-                                  'load-file #(load-file* sci-ctx %))))
-            _ (swap! (:env sci-ctx)
-                     (fn [env]
-                       (assoc-in env [:namespaces 'clojure.main 'repl]
-                                 (fn [& opts]
-                                   (let [opts (apply hash-map opts)]
-                                     (repl/start-repl! sci-ctx opts))))))
+                       (update env :namespaces
+                               (fn [namespaces] [:namespaces 'clojure.main 'repl]
+                                 (-> namespaces
+                                     (assoc-in ['clojure.core 'load-file] #(load-file* sci-ctx %))
+                                     (assoc-in ['clojure.main 'repl]
+                                               (fn [& opts]
+                                                 (let [opts (apply hash-map opts)]
+                                                   (repl/start-repl! sci-ctx opts))))
+                                     (assoc-in ['user (with-meta '*input*
+                                                        (when-not stream?
+                                                          {:sci.impl/deref! true}))] input-var))))))
             preloads (some-> (System/getenv "BABASHKA_PRELOADS") (str/trim))
             [expressions exit-code]
             (cond expressions [expressions nil]
@@ -395,17 +431,16 @@ Everything after that is bound to *command-line-args*."))
                        [(print-help) 0]
                        repl [(repl/start-repl! sci-ctx) 0]
                        socket-repl [(start-socket-repl! socket-repl sci-ctx) 0]
+                       nrepl [(start-nrepl! nrepl sci-ctx) 0]
                        (not (str/blank? expression))
                        (try
-                         (loop [in (read-next *in*)]
-                           (let [_ (swap! env update-in [:namespaces 'user]
-                                          assoc (with-meta '*input*
-                                                  (when-not stream?
-                                                    {:sci.impl/deref! true}))
-                                          (sci/new-dynamic-var '*input* in))]
+                         (loop []
+                           (let [in (read-next *in*)]
                              (if (identical? ::EOF in)
                                [nil 0] ;; done streaming
-                               (let [res [(let [res (eval-string* sci-ctx expression)]
+                               (let [res [(let [res
+                                                (sci/binding [input-var in]
+                                                  (eval-string* sci-ctx expression))]
                                             (when (some? res)
                                               (if-let [pr-f (cond shell-out println
                                                                   edn-out prn)]
@@ -416,7 +451,7 @@ Everything after that is bound to *command-line-args*."))
                                                   (pr-f res))
                                                 (prn res)))) 0]]
                                  (if stream?
-                                   (recur (read-next *in*))
+                                   (recur)
                                    res)))))
                          (catch Throwable e
                            (error-handler* e verbose?)))
@@ -440,7 +475,8 @@ Everything after that is bound to *command-line-args*."))
 (defn -main
   [& args]
   (if-let [dev-opts (System/getenv "BABASHKA_DEV")]
-    (let [{:keys [:n]} (edn/read-string dev-opts)
+    (let [{:keys [:n]} (if (= "true" dev-opts) {:n 1}
+                           (edn/read-string dev-opts))
           last-iteration (dec n)]
       (dotimes [i n]
         (if (< i last-iteration)